AP-2140 StatementOfCases controller
